crew/requirements.txt
crewai
crewai-tools
boto3
python-dotenv
tavily-python
bedrock-agentcore<=0.1.5
langchain-community
langchain-tavily
duckduckgo-search
ddgs
bedrock_agentcore_starter_toolkit
